Homework 8
I struggled uploading hw7 so hopefully this is correct. Admittedly, i am uploading on the web page as i couldn't figure out how to send the hw directly to the 'hw8' branch via PyCharm - i only had the option to send it to the master branch when i clicked 'commit and push' on PyCharm. Could i please have some help with this problem for future homeworks? :)