Add files via upload
I enjoyed practicing testing for this homework, and feel that i understood it. However, in the case of this homework all of my tests failed. This is because i struggled to re-write my ATM function in the form of small functions (my initial code was not split into functions). I think it might have been better to have an additional second question where the to-be-tested code is given to us already, to to ensure that we can do the 'testing' part independently of the 'de-tangle code into quite sophisticated and concise functions'. I know it is important to be able to de-tangle the code -- hopefully i will get there with a bit more practice! I hope i don't get marked down too much for this, as i did understand the testing.